requirements.txt
python-telegram-bot==20.6
pip install --upgrade python-telegram-bot